Create swapchain on vulkan backend (no recreation for resize)

+++ engine/src/renderer/vulkan/vulkan_backend.h
@@ -9,7 +9,7 @@ struct box_engine;
 b8 vulkan_renderer_backend_initialize(renderer_backend* backend, renderer_backend_config* config);
 void vulkan_renderer_backend_shutdown(renderer_backend* backend);
 
-void vulkan_renderer_backend_on_resized(renderer_backend* backend, u16 width, u16 height);
+void vulkan_renderer_backend_on_resized(renderer_backend* backend, u32 width, u32 height);
 
 b8 vulkan_renderer_backend_begin_frame(renderer_backend* backend, f32 delta_time);
 b8 vulkan_renderer_backend_end_frame(renderer_backend* backend);
